requirements.txt
allure_python_commons==2.13.5
pytest==8.3.3
Requests==2.32.3
